fix: separate mistral from mistral3
